@@ -20,7 +20,6 @@
     "react-dom": "^15.3.1",
     "react-redux": "^4.4.5",
     "react-router": "^2.7.0",
-    "rxjs": "^5.0.0-rc.1",
     "socket.io-client": "^1.5.1"
   },
   "devDependencies": {
@@ -35,7 +34,8 @@
     "gulp-autoprefixer": "^3.1.1",
     "gulp-nodemon": "^2.2.1",
     "gulp-sass": "^2.3.2",
-    "webpack": "^1.0.0",
+    "rxjs": "^5.0.3",
+    "webpack": "^2.2.0",
     "webpack-dev-server": "^1.16.2",
     "webpack-stream": "^3.2.0"
   }
