@@ -1,6 +1,7 @@
 import React from 'react';
+import locale from '../../config/locale';
 
 export default function Player({ name, rating, side }) {
-	return(<div>{`${name} ${rating ? `(${rating})` : ''} and side is ${side}`}</div>);
+	return(<div>{`${name} ${rating ? `(${rating})` : ''} ${locale.andSide} ${locale[side]}`}</div>);
 
 }
