@@ -1,9 +1,11 @@
 import React from 'react';
+import { Link } from 'react-router';
 
-export default function Match({ players }) {
+export default function Match({ players, roomId, type, time }) {
+	let stringedPlayers = players.map(player => `${player.name} (${player.rating})`);
 	return(
 		<li className="list-group-item">
-			{players[0]} - {players[1]}
+			<a href={`/checkers/match/${roomId}`}>{`${stringedPlayers[0]} - ${stringedPlayers[1]} | ${type}, ${time}`}</a>
 		</li>
 	);
 }
