@@ -1,38 +1,38 @@
-var x = {
-	"1-A":{
-		"id":'1-A',
-		"checker": 1,
+const x = {
+	'1-A': {
+		'id': '1-A',
+		'checker': 1,
 		connected: {
 			'2': '2-B'
 		}
 	},
-	"1-C":{
-		"id":'1-C',
-		"checker": 1,
+	'1-C': {
+		'id': '1-C',
+		'checker': 1,
 		connected: {
 			'1': '2-B',
 			'2': '2-D'
 		}
 	},
-	"1-E":{
-		"id":'1-E',
-		"checker": 1,
+	'1-E': {
+		'id': '1-E',
+		'checker': 1,
 		connected: {
 			'1': '2-D',
 			'2': '2-F'
 		}
 	},
-	"1-G":{
-		"id":'1-G',
-		"checker": 1,
+	'1-G': {
+		'id': '1-G',
+		'checker': 1,
 		connected: {
 			2: '2-H',
 			1: '2-F'
 		}
 	},
-	"2-B":{
-		"id":'2-B',
-		"checker": 1,
+	'2-B': {
+		'id': '2-B',
+		'checker': 1,
 		connected: {
 			1: '3-A',
 			2: '3-C',
@@ -40,9 +40,9 @@ var x = {
 			'-2': '1-C'
 		}
 	},
-	"2-D":{
-		"id":'2-D',
-		"checker": 1,
+	'2-D': {
+		'id': '2-D',
+		'checker': 1,
 		connected: {
 			2: '3-E',
 			1: '3-C',
@@ -50,9 +50,9 @@ var x = {
 			'-1': '1-C'
 		}
 	},
-	"2-F":{
-		"id":'2-F',
-		"checker": 1,
+	'2-F': {
+		'id': '2-F',
+		'checker': 1,
 		connected: {
 			1: '3-E',
 			2: '3-G',
@@ -60,25 +60,25 @@ var x = {
 			'-2': '1-G'
 		}
 	},
-	"2-H":{
-		"id":'2-H',
-		"checker": 1,
+	'2-H': {
+		'id': '2-H',
+		'checker': 1,
 		connected: {
 			1: '3-G',
 			'-1': '1-G'
 		}
 	},
-	"3-A":{
-		"id":'3-A',
-		"checker": 1,
+	'3-A': {
+		'id': '3-A',
+		'checker': 1,
 		connected: {
 			'-2': '2-B',
 			'2': '4-B'
 		}
 	},
-	"3-C":{
-		"id":'3-C',
-		"checker": 0,
+	'3-C': {
+		'id': '3-C',
+		'checker': 0,
 		connected: {
 			'-1': '2-B',
 			1: '4-B',
@@ -86,9 +86,9 @@ var x = {
 			'-2': '2-D'
 		}
 	},
-	"3-E":{
-		"id":'3-E',
-		"checker": 1,
+	'3-E': {
+		'id': '3-E',
+		'checker': 1,
 		connected: {
 			'-2': '2-F',
 			2: '4-F',
@@ -96,9 +96,9 @@ var x = {
 			'-1': '2-D'
 		}
 	},
-	"3-G":{
-		"id":'3-G',
-		"checker": 1,
+	'3-G': {
+		'id': '3-G',
+		'checker': 1,
 		connected: {
 			'-1': '2-F',
 			1: '4-F',
@@ -106,9 +106,9 @@ var x = {
 			'-2': '2-H'
 		}
 	},
-	"4-B":{
-		"id":'4-B',
-		"checker": 1,
+	'4-B': {
+		'id': '4-B',
+		'checker': 1,
 		connected: {
 			1: '5-A',
 			2: '5-C',
@@ -116,9 +116,9 @@ var x = {
 			'-2': '3-C'
 		}
 	},
-	"4-D":{
-		"id":'4-D',
-		"checker": 1,
+	'4-D': {
+		'id': '4-D',
+		'checker': 1,
 		connected: {
 			2: '5-E',
 			1: '5-C',
@@ -126,9 +126,9 @@ var x = {
 			'-1': '3-C'
 		}
 	},
-	"4-F":{
-		"id":'4-F',
-		"checker": 0,
+	'4-F': {
+		'id': '4-F',
+		'checker': 0,
 		connected: {
 			1: '5-E',
 			2: '5-G',
@@ -136,25 +136,25 @@ var x = {
 			'-2': '3-G'
 		}
 	},
-	"4-H":{
-		"id":'4-H',
-		"checker": 0,
+	'4-H': {
+		'id': '4-H',
+		'checker': 0,
 		connected: {
 			1: '5-G',
 			'-1': '3-G'
 		}
 	},
-	"5-A":{
-		"id":'5-A',
-		"checker": 0,
+	'5-A': {
+		'id': '5-A',
+		'checker': 0,
 		connected: {
 			2: '6-B',
 			'-2': '4-B'
 		}
 	},
-	"5-C":{
-		"id":'5-C',
-		"checker": 0,
+	'5-C': {
+		'id': '5-C',
+		'checker': 0,
 		connected: {
 			'1': '6-B',
 			'2': '6-D',
@@ -162,9 +162,9 @@ var x = {
 			'-2': '4-D'
 		}
 	},
-	"5-E":{
-		"id":'5-E',
-		"checker": -1,
+	'5-E': {
+		'id': '5-E',
+		'checker': -1,
 		connected: {
 			'2': '6-F',
 			'1': '6-D',
@@ -172,9 +172,9 @@ var x = {
 			'-1': '4-D'
 		}
 	},
-	"5-G":{
-		"id":'5-G',
-		"checker": -1,
+	'5-G': {
+		'id': '5-G',
+		'checker': -1,
 		connected: {
 			'1': '6-F',
 			'2': '6-H',
@@ -182,9 +182,9 @@ var x = {
 			'-2': '4-H'
 		}
 	},
-	"6-B":{
-		"id":'6-B',
-		"checker": -1,
+	'6-B': {
+		'id': '6-B',
+		'checker': -1,
 		connected: {
 			'1': '7-A',
 			'2': '7-C',
@@ -192,9 +192,9 @@ var x = {
 			'-2': '5-C'
 		}
 	},
-	"6-D":{
-		"id":'6-D',
-		"checker": -1,
+	'6-D': {
+		'id': '6-D',
+		'checker': -1,
 		connected: {
 			'1': '7-C',
 			'2': '7-E',
@@ -202,9 +202,9 @@ var x = {
 			'-2': '5-E'
 		}
 	},
-	"6-F":{
-		"id":'6-F',
-		"checker": 1,
+	'6-F': {
+		'id': '6-F',
+		'checker': 1,
 		connected: {
 			'1': '7-E',
 			'2': '7-G',
@@ -212,25 +212,25 @@ var x = {
 			'-2': '5-G'
 		}
 	},
-	"6-H":{
-		"id":'6-H',
-		"checker": -1,
+	'6-H': {
+		'id': '6-H',
+		'checker': -1,
 		connected: {
 			'1': '7-G',
 			'-1': '5-G'
 		}
 	},
-	"7-A":{
-		"id":'7-A',
-		"checker": -1,
+	'7-A': {
+		'id': '7-A',
+		'checker': -1,
 		connected: {
 			'2': '8-B',
 			'-2': '6-B'
 		}
 	},
-	"7-C":{
-		"id":'7-C',
-		"checker": -1,
+	'7-C': {
+		'id': '7-C',
+		'checker': -1,
 		connected: {
 			'1': '8-B',
 			'2': '8-D',
@@ -238,9 +238,9 @@ var x = {
 			'-2': '6-D'
 		}
 	},
-	"7-E":{
-		"id":'7-E',
-		"checker": -1,
+	'7-E': {
+		'id': '7-E',
+		'checker': -1,
 		connected: {
 			'2': '8-F',
 			'1': '8-D',
@@ -248,9 +248,9 @@ var x = {
 			'-1': '6-D'
 		}
 	},
-	"7-G":{
-		"id":'7-G',
-		"checker": 0,
+	'7-G': {
+		'id': '7-G',
+		'checker': 0,
 		connected: {
 			'1': '8-F',
 			'2': '8-H',
@@ -258,74 +258,74 @@ var x = {
 			'-2': '6-H'
 		}
 	},
-	"8-B":{
-		"id":'8-B',
-		"checker": -1,
+	'8-B': {
+		'id': '8-B',
+		'checker': -1,
 		connected: {
 			'-1': '7-A',
 			'-2': '7-C'
 		}
 	},
-	"8-D":{
-		"id":'8-D',
-		"checker": 0,
+	'8-D': {
+		'id': '8-D',
+		'checker': 0,
 		connected: {
 			'-2': '7-E',
 			'-1': '7-C'
 		}
 	},
-	"8-F":{
-		"id":'8-F',
-		"checker": -1,
+	'8-F': {
+		'id': '8-F',
+		'checker': -1,
 		connected: {
 			'-1': '7-E',
 			'-2': '7-G'
 		}
 	},
-	"8-H":{
-		"id":'8-H',
-		"checker": -1,
+	'8-H': {
+		'id': '8-H',
+		'checker': -1,
 		connected: {
 			'-1': '7-G'
 		}
 	}
-}
+};
 
 export default {
-	"1-A":{
-		"id":'1-A',
-		"checker": 1,
+	'1-A': {
+		'id': '1-A',
+		'checker': 1,
 		connected: {
 			'2': '2-B'
 		}
 	},
-	"1-C":{
-		"id":'1-C',
-		"checker": 1,
+	'1-C': {
+		'id': '1-C',
+		'checker': 1,
 		connected: {
 			'1': '2-B',
 			'2': '2-D'
 		}
 	},
-	"1-E":{
-		"id":'1-E',
-		"checker": 1,
+	'1-E': {
+		'id': '1-E',
+		'checker': 1,
 		connected: {
 			'1': '2-D',
 			'2': '2-F'
 		}
 	},
-	"1-G":{
-		"id":'1-G',
-		"checker": 1,
+	'1-G': {
+		'id': '1-G',
+		'checker': 1,
 		connected: {
 			2: '2-H',
 			1: '2-F'
 		}
 	},
-	"2-B":{
-		"id":'2-B',
-		"checker": 1,
+	'2-B': {
+		'id': '2-B',
+		'checker': 1,
 		connected: {
 			1: '3-A',
 			2: '3-C',
@@ -333,9 +333,9 @@ export default {
 			'-2': '1-C'
 		}
 	},
-	"2-D":{
-		"id":'2-D',
-		"checker": 1,
+	'2-D': {
+		'id': '2-D',
+		'checker': 1,
 		connected: {
 			2: '3-E',
 			1: '3-C',
@@ -343,9 +343,9 @@ export default {
 			'-1': '1-C'
 		}
 	},
-	"2-F":{
-		"id":'2-F',
-		"checker": 1,
+	'2-F': {
+		'id': '2-F',
+		'checker': 1,
 		connected: {
 			1: '3-E',
 			2: '3-G',
@@ -353,25 +353,25 @@ export default {
 			'-2': '1-G'
 		}
 	},
-	"2-H":{
-		"id":'2-H',
-		"checker": 1,
+	'2-H': {
+		'id': '2-H',
+		'checker': 1,
 		connected: {
 			1: '3-G',
 			'-1': '1-G'
 		}
 	},
-	"3-A":{
-		"id":'3-A',
-		"checker": 1,
+	'3-A': {
+		'id': '3-A',
+		'checker': 1,
 		connected: {
 			'-2': '2-B',
 			'2': '4-B'
 		}
 	},
-	"3-C":{
-		"id":'3-C',
-		"checker": 1,
+	'3-C': {
+		'id': '3-C',
+		'checker': 1,
 		connected: {
 			'-1': '2-B',
 			1: '4-B',
@@ -379,9 +379,9 @@ export default {
 			'-2': '2-D'
 		}
 	},
-	"3-E":{
-		"id":'3-E',
-		"checker": 1,
+	'3-E': {
+		'id': '3-E',
+		'checker': 1,
 		connected: {
 			'-2': '2-F',
 			2: '4-F',
@@ -389,9 +389,9 @@ export default {
 			'-1': '2-D'
 		}
 	},
-	"3-G":{
-		"id":'3-G',
-		"checker": 1,
+	'3-G': {
+		'id': '3-G',
+		'checker': 1,
 		connected: {
 			'-1': '2-F',
 			1: '4-F',
@@ -399,9 +399,9 @@ export default {
 			'-2': '2-H'
 		}
 	},
-	"4-B":{
-		"id":'4-B',
-		"checker": 0,
+	'4-B': {
+		'id': '4-B',
+		'checker': 0,
 		connected: {
 			1: '5-A',
 			2: '5-C',
@@ -409,9 +409,9 @@ export default {
 			'-2': '3-C'
 		}
 	},
-	"4-D":{
-		"id":'4-D',
-		"checker": 0,
+	'4-D': {
+		'id': '4-D',
+		'checker': 0,
 		connected: {
 			2: '5-E',
 			1: '5-C',
@@ -419,9 +419,9 @@ export default {
 			'-1': '3-C'
 		}
 	},
-	"4-F":{
-		"id":'4-F',
-		"checker": 0,
+	'4-F': {
+		'id': '4-F',
+		'checker': 0,
 		connected: {
 			1: '5-E',
 			2: '5-G',
@@ -429,25 +429,25 @@ export default {
 			'-2': '3-G'
 		}
 	},
-	"4-H":{
-		"id":'4-H',
-		"checker": 0,
+	'4-H': {
+		'id': '4-H',
+		'checker': 0,
 		connected: {
 			1: '5-G',
 			'-1': '3-G'
 		}
 	},
-	"5-A":{
-		"id":'5-A',
-		"checker": 0,
+	'5-A': {
+		'id': '5-A',
+		'checker': 0,
 		connected: {
 			2: '6-B',
 			'-2': '4-B'
 		}
 	},
-	"5-C":{
-		"id":'5-C',
-		"checker": 0,
+	'5-C': {
+		'id': '5-C',
+		'checker': 0,
 		connected: {
 			'1': '6-B',
 			'2': '6-D',
@@ -455,9 +455,9 @@ export default {
 			'-2': '4-D'
 		}
 	},
-	"5-E":{
-		"id":'5-E',
-		"checker": 0,
+	'5-E': {
+		'id': '5-E',
+		'checker': 0,
 		connected: {
 			'2': '6-F',
 			'1': '6-D',
@@ -465,9 +465,9 @@ export default {
 			'-1': '4-D'
 		}
 	},
-	"5-G":{
-		"id":'5-G',
-		"checker": 0,
+	'5-G': {
+		'id': '5-G',
+		'checker': 0,
 		connected: {
 			'1': '6-F',
 			'2': '6-H',
@@ -475,9 +475,9 @@ export default {
 			'-2': '4-H'
 		}
 	},
-	"6-B":{
-		"id":'6-B',
-		"checker": -1,
+	'6-B': {
+		'id': '6-B',
+		'checker': -1,
 		connected: {
 			'1': '7-A',
 			'2': '7-C',
@@ -485,9 +485,9 @@ export default {
 			'-2': '5-C'
 		}
 	},
-	"6-D":{
-		"id":'6-D',
-		"checker": -1,
+	'6-D': {
+		'id': '6-D',
+		'checker': -1,
 		connected: {
 			'1': '7-C',
 			'2': '7-E',
@@ -495,9 +495,9 @@ export default {
 			'-2': '5-E'
 		}
 	},
-	"6-F":{
-		"id":'6-F',
-		"checker": -1,
+	'6-F': {
+		'id': '6-F',
+		'checker': -1,
 		connected: {
 			'1': '7-E',
 			'2': '7-G',
@@ -505,25 +505,25 @@ export default {
 			'-2': '5-G'
 		}
 	},
-	"6-H":{
-		"id":'6-H',
-		"checker": -1,
+	'6-H': {
+		'id': '6-H',
+		'checker': -1,
 		connected: {
 			'1': '7-G',
 			'-1': '5-G'
 		}
 	},
-	"7-A":{
-		"id":'7-A',
-		"checker": -1,
+	'7-A': {
+		'id': '7-A',
+		'checker': -1,
 		connected: {
 			'2': '8-B',
 			'-2': '6-B'
 		}
 	},
-	"7-C":{
-		"id":'7-C',
-		"checker": -1,
+	'7-C': {
+		'id': '7-C',
+		'checker': -1,
 		connected: {
 			'1': '8-B',
 			'2': '8-D',
@@ -531,9 +531,9 @@ export default {
 			'-2': '6-D'
 		}
 	},
-	"7-E":{
-		"id":'7-E',
-		"checker": -1,
+	'7-E': {
+		'id': '7-E',
+		'checker': -1,
 		connected: {
 			'2': '8-F',
 			'1': '8-D',
@@ -541,9 +541,9 @@ export default {
 			'-1': '6-D'
 		}
 	},
-	"7-G":{
-		"id":'7-G',
-		"checker": -1,
+	'7-G': {
+		'id': '7-G',
+		'checker': -1,
 		connected: {
 			'1': '8-F',
 			'2': '8-H',
@@ -551,35 +551,35 @@ export default {
 			'-2': '6-H'
 		}
 	},
-	"8-B":{
-		"id":'8-B',
-		"checker": -1,
+	'8-B': {
+		'id': '8-B',
+		'checker': -1,
 		connected: {
 			'-1': '7-A',
 			'-2': '7-C'
 		}
 	},
-	"8-D":{
-		"id":'8-D',
-		"checker": -1,
+	'8-D': {
+		'id': '8-D',
+		'checker': -1,
 		connected: {
 			'-2': '7-E',
 			'-1': '7-C'
 		}
 	},
-	"8-F":{
-		"id":'8-F',
-		"checker": -1,
+	'8-F': {
+		'id': '8-F',
+		'checker': -1,
 		connected: {
 			'-1': '7-E',
 			'-2': '7-G'
 		}
 	},
-	"8-H":{
-		"id":'8-H',
-		"checker": -1,
+	'8-H': {
+		'id': '8-H',
+		'checker': -1,
 		connected: {
 			'-1': '7-G'
 		}
 	}
-}
+};
