@@ -2,20 +2,21 @@ const Room = require('./Room');
 
 function playHandler(data, store, emitter) {
 
-	const filteredRooms = store.filter(room => {
+	const appliableRooms = store.filter(room => {
 		return room.players.length < 2 &&
 			room.type === data.type &&
 			room.time === data.time;
 	});
-	console.log('play after filter', filteredRooms, store);
-	if (filteredRooms.length > 0) {
-		const id = Math.random() * (filteredRooms.length - 1);
-		emitter.emit('roomCreated', filteredRooms[id].toString());
+	console.log('[play] filtered rooms and whole store', appliableRooms, store);
+	if (appliableRooms.length > 0) {
+		const id = Math.random() * (appliableRooms.length - 1);
+		emitter.emit('roomCreated', appliableRooms[id].toString());
 	}
 	else
 		// check if this emitter (preferably ip) doesn't just spam server and let it create room
 		if (!emitter.room) {
 			const id = store.create(new Room(emitter, data, store.counter));
+			emitter.broadcast.emit('matches', store.toTransferenceProtocol());
 			emitter.emit('roomCreated', id);
 		}
 }
@@ -24,7 +25,7 @@ function enterHandler(data, store, emitter) {
 	if (store[data.id])
 		store[data.id].join(emitter, data.player);
 	else
-		emitter.emit('err', { type: 'No room with such id' });
+		emitter.emit('wrongRoom');
 }
 
 function moveHandler(data, store, emitter) {
